Chain File corrections and update
Some of the chain files I updated were done on an older chain file, and so some of the new information was lost.

I went through all chains that I found a github link for, verified the cosmos sdk version and change it respectively.

Some of the chains do not have a link to github.  Others have an improper link to their repo, some links go to another project within their repo and some were just wrong.

Probably worth figuring out how to auto fetch the current SDK version instead of it being a manual configuration.