Aura network
CORS has been fixed and is ready to go for when mainnet goes live.

After speaking with the team, they wanted to wait until next month for the listing on Pingpub.  They are on a 'testnet' right now in prep for Mainnet.  Current testnet is being treated as mainnet.

Sorry for the bouncing around.