Patch - Uptime incorrectly displaying down for chains below .45 and unable to resolve validator moniker in blocks
After fetch.js was changed there was issues where chains below `0.45` were showing as the whole network was RED.
In blocks, the HEX address was showing instead of the Validator Moniker.

I refactored the fetch.js to a ternary.
In uptime and blocks I implemented a version check. If it's less than `v0.45`, it will convert the hex address to base64, Otherwise it will use the prior method.

Tested in my development environment.  Uptimes are reported properly, showing missed blocks if there are any and monikers match with proposers.