Fix - Delegate/Redelegate decimal issue
The options for delegation and redelegation in the "Delegation" area for the Dashboard or Wallet details, when clicked only represented the amount with 2 decimals.  Often this was rounding up and when a user chooses to redelegate, if they enter in the amount shown can run into an issue where they cannot submit the transaction due to the share amount being incorrect.